+++ www/main/navigator/search/components/SearchContent.jsx
@@ -3,7 +3,16 @@ import { useStoreActions, useStoreState } from 'easy-peasy';
 import { remote, ipcRenderer } from 'electron';
 import banidb from '../../../common/constants/banidb';
 import { filters } from '../../utils';
-import { IconButton, InputBox, FilterDropdown, SearchResults } from '../../../common/sttm-ui';
+import { retrieveFilterOption } from '../utils';
+
+import { classNames } from '../../../common/utils';
+import {
+  IconButton,
+  InputBox,
+  FilterDropdown,
+  SearchResults,
+  FilterTag,
+} from '../../../common/sttm-ui';
 import { GurmukhiKeyboard } from './GurmukhiKeyboard';
 import { useNewShabad } from '../hooks/use-new-shabad';
 
@@ -22,6 +31,7 @@ const SearchContent = () => {
     searchQuery,
     currentSearchType,
     shortcuts,
+    searchShabadsCount,
   } = useStoreState(state => state.navigator);
   const {
     setCurrentWriter,
@@ -29,16 +39,22 @@ const SearchContent = () => {
     setCurrentSource,
     setSearchQuery,
     setShortcuts,
+    setSearchShabadsCount,
   } = useStoreActions(state => state.navigator);
 
   // Local State
   const [databaseProgress, setDatabaseProgress] = useState(1);
   const [query, setQuery] = useState('');
+  const [writerArray, setWriterArray] = useState([]);
+  const [raagArray, setRaagArray] = useState([]);
+  const [sourceArray, setSourceArray] = useState([]);
 
   const sourcesObj = banidb.SOURCE_TEXTS;
   const writersObj = banidb.WRITER_TEXTS;
   const raagsObj = banidb.RAAG_TEXTS;
 
+  const isShowFiltersTag =
+    currentWriter !== 'all' || currentRaag !== 'all' || currentSource !== 'all';
   // Gurmukhi Keyboard
   const [keyboardOpenStatus, setKeyboardOpenStatus] = useState(false);
   const HandleKeyboardToggle = () => {
@@ -63,7 +79,7 @@ const SearchContent = () => {
       : [];
   };
 
-  const [filteredShabads, setFilteredShabads] = useState([filters(mapVerseItems(searchData))]);
+  const [filteredShabads, setFilteredShabads] = useState([]);
 
   const openFirstResult = () => {
     if (searchQuery.length > 0 && filteredShabads.length > 0) {
@@ -85,9 +101,9 @@ const SearchContent = () => {
 
   useEffect(() => {
     setFilteredShabads(
-      filters(mapVerseItems(searchData), currentWriter, currentRaag, currentSource),
+      filters(mapVerseItems(searchData), currentWriter, currentRaag, writerArray, raagArray),
     );
-  }, [searchData, currentWriter, currentRaag, currentSource]);
+  }, [searchData, currentWriter, currentRaag]);
 
   // checks if keyboard shortcut is fired then it invokes the function
   useEffect(() => {
@@ -106,6 +122,12 @@ const SearchContent = () => {
     }
   }, [searchQuery]);
 
+  useEffect(() => {
+    if (searchShabadsCount !== filteredShabads.length) {
+      setSearchShabadsCount(filteredShabads.length);
+    }
+  }, [filteredShabads]);
+
   ipcRenderer.on('database-progress', data => {
     setDatabaseProgress(data.percent);
   });
@@ -121,13 +143,28 @@ const SearchContent = () => {
     };
   }, [query]);
 
+  useEffect(() => {
+    const wData = retrieveFilterOption(writersObj, 'writer');
+    wData.then(d => {
+      setWriterArray(d);
+    });
+    const rData = retrieveFilterOption(raagsObj, 'raag');
+    rData.then(d => {
+      setRaagArray(d);
+    });
+    const sData = retrieveFilterOption(sourcesObj, 'source');
+    sData.then(d => {
+      setSourceArray(d);
+    });
+  }, []);
+
   return (
     <div className="search-content-container">
       <div className="search-content">
         <InputBox
           placeholder={getPlaceholder()}
           disabled={databaseProgress < 1}
-          className={`${currentLanguage === 'gr' && 'gurmukhi'} mousetrap`}
+          className={`${currentLanguage === 'gr' ? 'gurmukhi' : 'english'} mousetrap`}
           databaseProgress={databaseProgress}
           query={query}
           setQuery={setQuery}
@@ -151,7 +188,31 @@ const SearchContent = () => {
         <GurmukhiKeyboard searchType={currentSearchType} query={query} setQuery={setQuery} />
       )}
       <div className="search-result-controls">
-        <span>{filteredShabads.length ? `${filteredShabads.length} Results` : ''}</span>
+        {isShowFiltersTag && (
+          <div className="filter-tag--container">
+            {currentWriter !== 'all' && (
+              <FilterTag
+                close={() => setCurrentWriter('all')}
+                title={currentWriter}
+                filterType={i18n.t('SEARCH.WRITER')}
+              />
+            )}
+            {currentRaag !== 'all' && (
+              <FilterTag
+                close={() => setCurrentRaag('all')}
+                title={currentRaag}
+                filterType={i18n.t('SEARCH.RAAG')}
+              />
+            )}
+            {currentSource !== 'all' && (
+              <FilterTag
+                close={() => setCurrentSource('all')}
+                title={i18n.t(`SEARCH.SOURCES.${sourcesObj[currentSource]}.TEXT`)}
+                filterType={i18n.t('SEARCH.SOURCE')}
+              />
+            )}
+          </div>
+        )}
         <div className="filters">
           <span>Filter by </span>
           <FilterDropdown
@@ -160,7 +221,8 @@ const SearchContent = () => {
               setCurrentWriter(event.target.value);
               analytics.trackEvent('search', 'searchWriter', event.target.value);
             }}
-            optionsObj={writersObj}
+            optionsArray={writerArray}
+            currentValue={currentWriter}
           />
           <FilterDropdown
             title="Raag"
@@ -168,7 +230,8 @@ const SearchContent = () => {
               setCurrentRaag(event.target.value);
               analytics.trackEvent('search', 'searchRaag', event.target.value);
             }}
-            optionsObj={raagsObj}
+            optionsArray={raagArray}
+            currentValue={currentRaag}
           />
           <FilterDropdown
             title="Source"
@@ -176,11 +239,12 @@ const SearchContent = () => {
               setCurrentSource(event.target.value);
               analytics.trackEvent('search', 'searchSource', event.target.value);
             }}
-            optionsObj={sourcesObj}
+            optionsArray={sourceArray}
+            currentValue={currentSource}
           />
         </div>
       </div>
-      <div className="search-results">
+      <div className={classNames('search-results', isShowFiltersTag && 'filter-applied')}>
         <div className="verse-block">
           {filteredShabads.map(
             ({ ang, shabadId, sourceId, verse, verseId, writer, raag }, index) => (